requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
Faker==4.0.1
idna==2.8
Markdown==3.2.1
mysqlclient==1.4.6
Pillow==7.0.0
pyTelegramBotAPI==3.6.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.8
